tests: Refactor shared_object options into submodule
